format

+++ lumino/LuminoCore/include/LuminoCore/Base/StringFormat.hpp
@@ -4,14 +4,18 @@
 
 namespace ln {
 
-/** 複合書式文字列と可変長引数リストから文字列を生成します。 */
+/**
+ * 複合書式文字列と可変長引数リストから文字列を生成します。
+ * 
+ * 書式は std::fmt に従います。
+ * https://fmt.dev/latest/index.html
+ */
 template<typename... TArgs>
-inline String format(const StringRef& format, TArgs&&... args)
-{
+inline String format(const StringRef& format, TArgs&&... args) {
     try {
         std::basic_string_view<Char> view(format.data(), format.length());
         auto str = ::fmt::format(view, std::forward<TArgs>(args)...);
-        return String(str.c_str(), str.length());
+        return String(str.c_str(), static_cast<int>(str.length()));
     }
     catch (::fmt::format_error& e) {
         LN_ERROR("String::format errro: {}", e.what());
